requirements.txt
aliyun-python-sdk-core==2.13.30
aliyun-python-sdk-core-v3==2.13.11
aliyun-python-sdk-kms==2.14.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
crcmod==1.7
cryptography==3.2.1
idna==2.10
jmespath==0.10.0
oss2==2.13.1
pycparser==2.20
pycryptodome==3.9.9
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
urllib3==1.26.2
wincertstore==0.2
